refactor of invite/register spec for additional coverage
