+++ app/models/stripe_wrapper.rb
@@ -2,17 +2,17 @@ module StripeWrapper
   class Charge
     attr_accessor :response, :status
 
-    def initialize(response, status)
+    def initialize(response:, status:)
       self.response = response
       self.status = status
     end
 
-    def self.create(options={})
+    def self.create(amount:, source:, description:)
       begin
-        response = Stripe::Charge.create currency: 'usd', amount: options[:amount], source: options[:token], description: options[:description]
-        new response, :success
+        response = Stripe::Charge.create currency: 'usd', amount: amount, source: source, description: description
+        new response: response, status: :success
       rescue Stripe::CardError => error
-        new error, :error
+        new response: error, status: :error
       end
     end
