+++ spec/cassettes/A_new_user_registers_for_myflix/from_an_invitation_email/user_invites_a_friend_with_an_unregistered_email.yml
@@ -1,0 +1,124 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.stripe.com/v1/charges
+    body:
+      encoding: US-ASCII
+      string: amount=999&currency=usd&description=First-month+subscription+fee+for+Adeline+O%27Connell&source=tok_17xjVEH5BD0l29UeJtIUl2jy
+    headers:
+      Accept:
+      - "*/*; q=0.5, application/xml"
+      Accept-Encoding:
+      - gzip, deflate
+      User-Agent:
+      - Stripe/v1 RubyBindings/1.39.0
+      Authorization:
+      - Bearer sk_test_gjrjS85XDYH5zREDVUmXk2kQ
+      Content-Type:
+      - application/x-www-form-urlencoded
+      X-Stripe-Client-User-Agent:
+      - '{"bindings_version":"1.39.0","lang":"ruby","lang_version":"2.1.2 p95 (2014-05-08)","platform":"x86_64-darwin13.0","engine":"ruby","publisher":"stripe","uname":"Darwin
+        Aarons-MacBook-Pro.local 15.4.0 Darwin Kernel Version 15.4.0: Fri Feb 26 22:08:05
+        PST 2016; root:xnu-3248.40.184~3/RELEASE_X86_64 x86_64","hostname":"Aarons-MacBook-Pro.local"}'
+      Content-Length:
+      - '124'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Server:
+      - nginx
+      Date:
+      - Thu, 07 Apr 2016 20:13:20 GMT
+      Content-Type:
+      - application/json
+      Content-Length:
+      - '1530'
+      Connection:
+      - keep-alive
+      Access-Control-Allow-Credentials:
+      - 'true'
+      Access-Control-Allow-Methods:
+      - GET, POST, HEAD, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Max-Age:
+      - '300'
+      Cache-Control:
+      - no-cache, no-store
+      Request-Id:
+      - req_8EBtPm9e9YYL4m
+      Stripe-Version:
+      - '2016-03-07'
+      Strict-Transport-Security:
+      - max-age=31556926; includeSubDomains
+    body:
+      encoding: UTF-8
+      string: |
+        {
+          "id": "ch_17xjVEH5BD0l29UeHgtyzuWG",
+          "object": "charge",
+          "amount": 999,
+          "amount_refunded": 0,
+          "application_fee": null,
+          "balance_transaction": "txn_17xjVEH5BD0l29UeQRS1CKMu",
+          "captured": true,
+          "created": 1460060000,
+          "currency": "usd",
+          "customer": null,
+          "description": "First-month subscription fee for Adeline O'Connell",
+          "destination": null,
+          "dispute": null,
+          "failure_code": null,
+          "failure_message": null,
+          "fraud_details": {},
+          "invoice": null,
+          "livemode": false,
+          "metadata": {},
+          "order": null,
+          "paid": true,
+          "receipt_email": null,
+          "receipt_number": null,
+          "refunded": false,
+          "refunds": {
+            "object": "list",
+            "data": [],
+            "has_more": false,
+            "total_count": 0,
+            "url": "/v1/charges/ch_17xjVEH5BD0l29UeHgtyzuWG/refunds"
+          },
+          "shipping": null,
+          "source": {
+            "id": "card_17xjVEH5BD0l29Ue1Ldptdoz",
+            "object": "card",
+            "address_city": null,
+            "address_country": null,
+            "address_line1": null,
+            "address_line1_check": null,
+            "address_line2": null,
+            "address_state": null,
+            "address_zip": null,
+            "address_zip_check": null,
+            "brand": "Visa",
+            "country": "US",
+            "customer": null,
+            "cvc_check": "pass",
+            "dynamic_last4": null,
+            "exp_month": 1,
+            "exp_year": 2020,
+            "fingerprint": "ftJsWM0BXmH2mbMj",
+            "funding": "credit",
+            "last4": "4242",
+            "metadata": {},
+            "name": "Adeline O'Connell",
+            "tokenization_method": null
+          },
+          "source_transfer": null,
+          "statement_descriptor": null,
+          "status": "succeeded"
+        }
+    http_version: 
+  recorded_at: Thu, 07 Apr 2016 20:13:21 GMT
+recorded_with: VCR 3.0.1
